+++ packages/winnow/src/geometry/project-coordinates.js
@@ -1,18 +1,12 @@
 const proj4 = require('proj4')
+const transformCoordinates = require('./transform-coordinates')
 
 module.exports = function projectCoordinates (coordinates, options = {}) {
   const inSR = options.inSR || 'EPSG:4326'
   const outSR = options.outSR || 'EPSG:3857'
   if (inSR === outSR) return coordinates
-  if (Array.isArray(coordinates[0]) && Array.isArray(coordinates[0][0])) {
-    return coordinates.map(a => {
-      return projectCoordinates(a, options)
-    })
-  } else if (Array.isArray(coordinates[0]) && typeof coordinates[0][0] === 'number') {
-    return coordinates.map(a => {
-      return projectCoordinates(a, options)
-    })
-  } else {
-    return proj4(inSR, outSR, coordinates)
-  }
+
+  return transformCoordinates(coordinates, { inSR, outSR }, (coordinates, options) => {
+    return proj4(options.inSR, options.outSR, coordinates)
+  })
 }