+++ packages/winnow/src/geometry/reduce-precision.js
@@ -1,0 +1,8 @@
+const transformCoordinates = require('./transform-coordinates')
+module.exports = function (coordinates, precision) {
+  return transformCoordinates(coordinates, precision, (coordinates, precision) => {
+    return coordinates.map(position => {
+      return parseFloat(position.toFixed(precision))
+    })
+  })
+}