+++ packages/winnow/test/project.js
@@ -16,6 +16,18 @@ test('Project to Web Mercator using 3857', t => {
   t.equal(results[0].geometry.spatialReference, undefined)
 })
 
+test('Project to Web Mercator using 3857 and reduce the precision', t => {
+  t.plan(2)
+  const options = {
+    projection: 3857,
+    limit: 1,
+    geometryPrecision: 3
+  }
+  const results = winnow.query(features, options)
+  t.deepEqual(results[0].geometry.coordinates, [-11682713.392, 4857924.005])
+  t.equal(results[0].geometry.spatialReference, undefined)
+})
+
 test('Project to Web Mercator using 3857 and translating to esri', t => {
   t.plan(3)
   const options = {